Can now take MCA spectra from EmulatedInterface
MCA.h
    - Shortened the StoreData method parameters to short (same updates in MCA.cpp)
    - Updated the ADC_SIZE to match Pixie-16 maximum bit resolution

MCA_ROOT
    - Changed from TH1F to TH1D so we can read histograms with rootHistogramViewer
    - Updated style
    - CalculateHistogramId dries out some calls
    - moved all pointer initialization to nullptr.

+++ Acquisition/MCA/source/MCA.cpp
@@ -77,8 +77,8 @@ bool MCA::Step() {
     if (!pif_ || !pif_->CheckRunStatus())
         return false;
 
-    for (int mod = 0; mod < pif_->GetConfiguration().GetNumberOfModules(); mod++)
-        for (unsigned int ch = 0; ch < pif_->GetConfiguration().GetNumberOfChannels(); ch++)
+    for (unsigned short mod = 0; mod < pif_->GetConfiguration().GetNumberOfModules(); mod++)
+        for (unsigned short ch = 0; ch < pif_->GetConfiguration().GetNumberOfChannels(); ch++)
             StoreData(mod, ch);
 
     Flush();
@@ -88,4 +88,4 @@ bool MCA::Step() {
     return true;
 }
 
-bool MCA::StoreData(const int &mod, const int &ch) { return true; }
+bool MCA::StoreData(const unsigned short &mod, const unsigned short &ch) { return true; }
